Fixed gst-plugins-bad * Made optional dependences features * Added webrtc feature

+++ ports/gstreamer/portfile.cmake
@@ -65,8 +65,6 @@ else()
     set(PLUGIN_BASE_SUPPORT disabled)
 endif()
 if("plugins-bad" IN_LIST FEATURES)
-    # requires 'libdrm', 'dssim', 'libmicrodns'
-    message(FATAL_ERROR "The feature 'plugins-bad' is not supported in this port version")
     set(PLUGIN_BAD_SUPPORT enabled)
 else()
     set(PLUGIN_BAD_SUPPORT disabled)
@@ -81,6 +79,18 @@ if ("gl-graphene" IN_LIST FEATURES)
     set(GL_GRAPHENE enabled)
 else()
     set(GL_GRAPHENE disabled)
+endif()
+
+if ("webrtc" IN_LIST FEATURES)
+    set(PLUGIN_BAD_WEBRTC enabled)
+else()
+    set(PLUGIN_BAD_WEBRTC disabled)
+endif()
+
+if ("videoparsers" IN_LIST FEATURES)
+    set(PLUGIN_BAD_VIDEOPARSERS enabled)
+else()
+    set(PLUGIN_BAD_VIDEOPARSERS disabled)
 endif()
 
 if ("flac" IN_LIST FEATURES)
@@ -184,11 +194,13 @@ vcpkg_configure_meson(
         -Dgst-plugins-bad:default_library=${LIBRARY_LINKAGE}
         -Dgst-plugins-bad:opencv=disabled
         -Dgst-plugins-bad:hls-crypto=openssl
+        -Dgst-plugins-bad:webrtc=${PLUGIN_BAD_WEBRTC}
+        -Dgst-plugins-bad:videoparsers=${PLUGIN_BAD_VIDEOPARSERS}
         -Dgst-plugins-bad:examples=disabled
         -Dgst-plugins-bad:tests=disabled
         -Dgst-plugins-bad:doc=disabled
         -Dgst-plugins-bad:introspection=disabled
-        -Dgst-plugins-bad:nls=${LIBRARY_LINKAGE}
+        -Dgst-plugins-bad:nls=disabled
         -Dgst-plugins-bad:orc=disabled
         # gst-plugins-ugly
         -Dugly=${PLUGIN_UGLY_SUPPORT}
@@ -242,22 +254,21 @@ vcpkg_configure_meson(
 vcpkg_install_meson()
 
 # Remove duplicated GL headers (we already have `opengl-registry`)
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/include/KHR"
-                    "${CURRENT_PACKAGES_DIR}/include/GL"
-)
-if(NOT VCPKG_TARGET_IS_LINUX)
+if (GL_GRAPHENE STREQUAL "enabled")
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/include/KHR"
+                        "${CURRENT_PACKAGES_DIR}/include/GL"
+    )
     file(RENAME "${CURRENT_PACKAGES_DIR}/lib/gstreamer-1.0/include/gst/gl/gstglconfig.h"
                 "${CURRENT_PACKAGES_DIR}/include/gstreamer-1.0/gst/gl/gstglconfig.h"
     )
-endif()
-
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/share"
-                    "${CURRENT_PACKAGES_DIR}/debug/libexec"
-                    "${CURRENT_PACKAGES_DIR}/debug/lib/gstreamer-1.0/include"
-                    "${CURRENT_PACKAGES_DIR}/libexec"
-                    "${CURRENT_PACKAGES_DIR}/lib/gstreamer-1.0/include"
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/share"
+                        "${CURRENT_PACKAGES_DIR}/debug/libexec"
+                        "${CURRENT_PACKAGES_DIR}/debug/lib/gstreamer-1.0/include"
+                        "${CURRENT_PACKAGES_DIR}/libexec"
+                        "${CURRENT_PACKAGES_DIR}/lib/gstreamer-1.0/include"
                     "${CURRENT_PACKAGES_DIR}/share/gdb"
-)
+    )
+endif ()
 
 if(VCPKG_LIBRARY_LINKAGE STREQUAL "static")
     file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/bin"
